// src/common/exception/enum/error-message.enum.ts
export enum ErrorMessage {
    EMAIL_VALIDATE = 'Email Validate Failed',
    PASSWORD_VALIDATE = 'Password Validate Failed',
    LOGIN_FAILED = 'Something Error Id Or Password',
    DUPLICATED_PASSWORD = 'Duplicated Password',
    DUPLICATED_EMAIL = 'Duplicated Email',
    LOGIN_COUNT_EXCEED = 'Login Try Count Exceed'
}